+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Applied Energistics Plugin
     - Add in world fluid and seed crafting alongside condenser singularity and matter_ball
 - Moved Util input handling code to IngredientHandler
+- Config category uses getCategory function
 
 ### Removed
 - Harvestcraft jar, we fetch from jitpack now